fix(core): compute rolling-checksum blockSize*remove in 64-bit
RollingChecksum.Roll computed blockSize * remove as int32. blockSize is the
chunk size and remove is a byte (≤255), so the product overflows int32 once
chunkSize exceeds ~8.42 MB, producing a wrong weak checksum. The strong-hash
resolution means this never corrupts patched data, but it silently disables
rsync block matching for large chunks (every window misses → full refetch).
Widen the multiplication to long.

+++ Hina.Core/Rsync/RollingChecksum.cs
@@ -25,7 +25,10 @@ namespace Hina.Core.Rsync
             uint b = (prev >> 16) & 0xFFFF;
 
             a = (a + Mod - remove + add) % Mod;
-            b = (b + Mod - (uint)(blockSize * remove % Mod) + a) % Mod;
+            // blockSize * remove must be computed in 64-bit: blockSize is the chunk size and
+            // `remove` can be up to 255, so the product overflows int32 once chunkSize exceeds
+            // ~8.4 MB, corrupting the weak checksum (and silently disabling rsync matching).
+            b = (b + Mod - (uint)((long)blockSize * remove % Mod) + a) % Mod;
 
             return (b << 16) | a;
         }
